requirements.txt
tensorflow==2.11.0
streamlit==1.14.0
opencv-python==4.6.0.66
numpy==1.23.0
Pillow==9.0.1
matplotlib==3.5.1
scikit-learn==1.1.2
